+++ internal/integration/gomega_test.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/derision-test/go-mockgen/v2/internal/integration/testdata"
-	"github.com/derision-test/go-mockgen/v2/internal/integration/testdata/mocks"
-	. "github.com/derision-test/go-mockgen/v2/testutil/gomega"
 	. "github.com/onsi/gomega"
+	"github.com/unknwon/go-mockgen/internal/integration/testdata"
+	"github.com/unknwon/go-mockgen/internal/integration/testdata/mocks"
+	. "github.com/unknwon/go-mockgen/testutil/gomega"
 )
 
 func TestGomegaCalls(t *testing.T) {