+++ internal/mockgen/consts/consts.go
@@ -2,7 +2,7 @@ package consts
 
 const (
 	Name        = "go-mockgen"
-	PackageName = "github.com/derision-test/go-mockgen/v2"
+	PackageName = "github.com/unknwon/go-mockgen"
 	Description = "go-mockgen generates mock implementations from interface definitions."
 	Version     = "2.1.1"
 )