+++ testutil/gomega/called_matcher.go
@@ -3,9 +3,9 @@ package matchers
 import (
 	"fmt"
 
-	"github.com/derision-test/go-mockgen/v2/internal/testutil"
 	"github.com/onsi/gomega/format"
 	"github.com/onsi/gomega/types"
+	"github.com/unknwon/go-mockgen/internal/testutil"
 )
 
 type calledMatcher struct {