+++ testutil/gomega/called_with_matcher.go
@@ -3,10 +3,10 @@ package matchers
 import (
 	"fmt"
 
-	"github.com/derision-test/go-mockgen/v2/internal/testutil"
 	"github.com/onsi/gomega/format"
 	"github.com/onsi/gomega/matchers"
 	"github.com/onsi/gomega/types"
+	"github.com/unknwon/go-mockgen/internal/testutil"
 )
 
 type calledWithMatcher struct {